// src/modules/properties/interfaces/IFilters.ts
export interface IFilters {
  city: string
  zone: string
  orderBy: string
  priceMin: string
  priceMax: string
  sizeMin: string
  sizeMax: string
  rooms: string
  baths: string
  nosell: boolean
  norent: boolean
  notransfer: boolean
  terraza: boolean
  ascensor: boolean
  tipoOfer: string[]
}